Modifications to build under Xcode 4, iOS SDK 5

+++ Classes/LogViewController.m
@@ -254,7 +254,7 @@ static NSString * const kHideBadgeKey = @"LogViewControllerHideBadge";
 
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-	if ((NSUInteger)section == [lastIndexPath section]) {
+	if (section == [lastIndexPath section]) {
 		return [lastIndexPath row] + 1;
 	} else {
 		EWMonth month = [self monthForSection:section];
